+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Show only `Done` button on macOS
+
 ## 1.1.1 - 2022 Dec 10
 - Update dependencies